app/bot/requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
annotated-types==0.6.0
certifi==2023.11.17
flake8==7.0.0
magic-filter==1.0.12
mccabe==0.7.0
pycodestyle==2.11.1
pydantic==1.10.13
pydantic_core==2.14.6
pyflakes==3.2.0
Pygments==2.17.2
python-dotenv==1.0.0
